project structure overhaul

+++ scripts/release.sh
@@ -5,6 +5,7 @@ set -e -o pipefail
 source "$(dirname "${BASH_SOURCE[0]}")/_config.sh"
 
 TASK=${1:-compile-dirac-pseudo-names}
+INCREMENTAL=${DIRAC_INCREMENTAL_RELEASE}
 
 TMP_RELEASE_BUILD="$DEPOT_DIR/out/Default/resources/inspector"
 
@@ -12,29 +13,33 @@ cd "$ROOT"
 
 set -x
 
-"$SCRIPTS/clean-target.sh"
-"$SCRIPTS/check-versions.sh"
-"$SCRIPTS/depot-sync.sh"
-"$SCRIPTS/depot-clean.sh"
+if [[ -z "$INCREMENTAL" ]]; then
+  "$SCRIPTS/check-versions.sh"
+  "$SCRIPTS/clean-target.sh"
+  "$SCRIPTS/depot-clean.sh"
+  "$SCRIPTS/depot-sync.sh"
+fi
 
-SHELL_JS="$RELEASE_BUILD/devtools/front_end/shell.js"
+IMPLANTED_JS="$RELEASE_BUILD/devtools/front_end/inspector.js"
 
-# make sure we rebuild shell.js from scratch
-if [[ -f "$SHELL_JS" ]]; then
-  rm "$SHELL_JS"
+# make sure we rebuild the js file from scratch
+if [[ -f "$IMPLANTED_JS" ]]; then
+  rm "$IMPLANTED_JS"
 fi
 "$SCRIPTS/depot-build-devtools.sh" "$TASK"
 
-rm -rf "$RELEASE_BUILD_DEVTOOLS_FRONTEND"
+if [[ -z "$DIRAC_INCREMENTAL_RELEASE" ]]; then
+  rm -rf "$RELEASE_BUILD_DEVTOOLS_FRONTEND"
+fi
 mkdir -p "$RELEASE_BUILD_DEVTOOLS_FRONTEND"
-cp -R "$TMP_RELEASE_BUILD/." "$RELEASE_BUILD_DEVTOOLS_FRONTEND"
+rsync -a "$TMP_RELEASE_BUILD/." "$RELEASE_BUILD_DEVTOOLS_FRONTEND"
 
 # copy compiled extension code (produced by `lein compile-dirac`)
 cp "$ROOT/target/resources/release/.compiled/background.js" "$RELEASE_BUILD/background.js"
 cp "$ROOT/target/resources/release/.compiled/options.js" "$RELEASE_BUILD/options.js"
 
 # concat bundled implant.js with shell.js
-echo ";/* DIRAC IMPLANT >> */" >>"$SHELL_JS"
-cat "$ROOT/target/resources/release/devtools/front_end/dirac/.compiled/implant/implant.js" >>"$SHELL_JS"
+echo ";/* DIRAC IMPLANT >> */" >>"$IMPLANTED_JS"
+cat "$ROOT/target/resources/release/devtools/front_end/dirac/.compiled/implant/implant.js" >>"$IMPLANTED_JS"
 
 "$SCRIPTS/prune-release.sh"
